+++ main/http_server/http_websocket.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "esp_http_server.h"
+
+esp_err_t echo_handler(httpd_req_t *req);
+void websocket_log_handler(void* param);
+
+void websocket_start();
